// src/shared/api/base/generated-types.ag.ts
export enum PSupportedLocaleIds {
	de_DE = 'de-DE',
	de_AT = 'de-AT',
	de_CH = 'de-CH',
	en_NL = 'en-NL',
	en_BE = 'en-BE',
	en_GB = 'en-GB',
	en_CZ = 'en-CZ',
	en_SE = 'en-SE',
	de_LU = 'de-LU',
	en = 'en',
}
// ...
export enum PSupportedCurrencyCodes {
	EUR = 'EUR',
	CHF = 'CHF',
	GBP = 'GBP',
	CZK = 'CZK',
	SEK = 'SEK',
}